% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoff_quantile.R
\name{ecoff_quantile}
\alias{ecoff_quantile}
\title{ECOFF Based On Quantile}
\usage{
ecoff_quantile(my_k, res, quant = 0.01)
}
\arguments{
\item{my_k}{number of components of the GMM, can be calculated using optimal_K(data, round(sqrt(sum(binned_data!=0))), "BIC", take_resist = F)}

\item{res}{result from the function EM_fit, can be calculated using EM_fit(binned_data, my_k, take_resist = F)}

\item{quant}{quantile to use as a decision for the ECOFF value, default is 0.01}
}
\value{
\itemize{
 \item{out: ECOFF}
}
}
\description{
Finds ECOFF value based on the quantile of the GMM chosen by some information criterion.
}
