% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_fit.R
\name{EM_fit}
\alias{EM_fit}
\title{EM Fit Of K-mixture Components}
\usage{
EM_fit(data, K, take_resist = F, plot = F, tolerance = 0.001)
}
\arguments{
\item{data}{vector of binned data}

\item{K}{number of gaussian components for mixture models - at least two}

\item{take_resist}{logical, whether or not to consider first bin refering to resisant observations, default is F}

\item{plot}{logical, whether or not to plot final fit of GMM on original data, default is F}

\item{tolerance}{tolerance for the Aitken-acceleration based stopping criterion, default is 0.001}
}
\value{
\itemize{
 \item {pi: mixing parameters pik of final GMM}
 \item {mu: means muk of the final GMM}
 \item {sigma: standard deviation sigma of the final homoscedastic GMM}
 \item {loglik: vector of log-likelihood during the iterative procedure - in case of K=1 only last value of log-likelihood will be returned}
 \item {AIC: AIC of the GMM fitted to the given data}
 \item {BIC: BIC of the GMM fitted to the given data}
 \item {data_used: data which were used for the fitting procedure}
 \item {reconstr_data: reconstructed data of the given binned data}
 \item {bins: bins}
}
}
\description{
For the estimation of GMM with only K=1 component we have used the function MASS::fitdistr().For the estimation of GMM with k=2,...,K components an EM algorithm is used to obtain needed output.
}
