% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_fun.R
\name{fitting_fun}
\alias{fitting_fun}
\title{Final ECOFF Value}
\usage{
fitting_fun(data, fit_one_mode = NULL, quant = 0.01, take_resist = F,
  plotting = F)
}
\arguments{
\item{data}{data which is used for analysis (one row of the data set)}

\item{fit_one_mode}{logical, whether function ecoff quantile should be called for the data, if they can not be clustered to 2 clusters by semi_param_clust, default is F}

\item{quant}{quantile which should be used for the function ecoff_quantile, default is 0.01}

\item{take_resist}{logical, whether the resistant - in our case Z6, i.e. the first column - should be used for the fitting and further procedure, default is F}

\item{plotting}{logical, whether the histogram of the data with fitted GMM and ECOFF value should be plotted, default is F}
}
\value{
\itemize{
 \item{my_ecoff: ECOFF}
}
}
\description{
As the ECOFF value can not always be calculated through semi-parametric clustering, this function decides how the ECOFF should be calculated (semi_param_clust or ecoff_quantile) and returns the final ECOFF. If there is not enough information in the data, this function will return ECOFF to be 0. This can happen in the case when we have less than 5 observations(i.e. 4 and less).
}
