% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_K.R
\name{optimal_K}
\alias{optimal_K}
\title{optimal_K}
\usage{
optimal_K(data, max_comp, criterion, take_resist = F)
}
\arguments{
\item{data}{binned data}

\item{max_comp}{the maximal K which should be tried, function will try 1:max comp values of}

\item{criterion}{which criterion should be used, one can choose either AIC or BIC}

\item{take_resist}{logical, whether resistant observations - first column of the data - should be used, default is FALSE}
}
\value{
\itemize{
 \item{opt: optimal K - number of components which minimize chosen information criterion for the given data set}
 }
}
\description{
Assesses the optimal number of components GMM.
}
