% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdata.R
\name{plotdata}
\alias{plotdata}
\title{Data Plots}
\usage{
plotdata(ZD, anti, bac)
}
\arguments{
\item{ZD}{data}

\item{anti}{Antimicrobial}

\item{bac}{Bacterium}
}
\description{
Plots given data.
}
\examples{
plotdata(ZD, "Ampicillin", "Escherichia coli")
plotdata(ZD, "Piperacillin", "Escherichia coli")
plotdata(ZD, "Mecillinam", "Escherichia coli")
}
