% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdens.R
\name{plotdens}
\alias{plotdens}
\title{EM Density And ECOFF Value}
\usage{
plotdens(ZD, ant, bac)
}
\arguments{
\item{ZD}{data}

\item{ant}{Antimicrobial}

\item{bac}{Bacterium}
}
\value{
plot of EM density and ECOFF value
}
\description{
Plots the EM density and the ECOFF Value. Additionally it gives information on the number of components.
}
\examples{
plotdens(ZD, "Ampicillin", "Escherichia coli")
plotdens(ZD, "Piperacillin", "Escherichia coli")
plotdens(ZD, "Mecillinam", "Escherichia coli")
}
