% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semi_param_clust.R
\name{semi_param_clust}
\alias{semi_param_clust}
\title{Semi-Parametric Clustering According To Scrucca}
\usage{
semi_param_clust(res, my_k, plot = F)
}
\arguments{
\item{res}{result from the function EM fit}

\item{my_k}{number of components of the GMM}

\item{plot}{if TRUE, the function will return a result nd plot the fitted density and the mode of the function - set FALSE by default}
}
\value{
\itemize{
 \item{Cluster1}
 \item{Cluster2}
 \item{ecoff}
 }
}
\description{
Uses Semi-Parametric Clustering to part the data into two clusters. The border between those two clusters is estimated to be the ECOFF.
}
\references{
Scrucca (2016) <doi:10.1016/j.csda.2015.01.006>
}
