% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testecoff.R
\name{testecoff}
\alias{testecoff}
\title{Validation Of ECOFF}
\usage{
testecoff(ZD)
}
\arguments{
\item{ZD}{data}
}
\value{
\itemize{
 \item{errors: number of ecoffs that could not be calculated}
 \item{RMSE: root mean squared error}
 }
number of ecoffs that could not be calculated, RMSE
}
\description{
Tests, if ECOFF can be calculated for all observed data points. Additionally it compares the ECOFF returned by fitting_fun to the official ECOFF values using a RSME.
}
