% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testecoffs.R
\name{testecoffs}
\alias{testecoffs}
\title{Validation Of ECOFF Using Only Scrucca's Model}
\usage{
testecoffs(ZD)
}
\arguments{
\item{ZD}{data}
}
\value{
\itemize{
 \item{errors_scrucca: number of ecoffs of observations with more than 4 components that could not be calculated}
 \item{RMSE_scrucca: root squared error}
}
}
\description{
Tests, if ECOFF can be calculated for all observed data points using only Scrucca's model. Additionally it compares the ECOFF returned by fitting_fun to the official ECOFF values using a RSME.
}
