% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZD}
\alias{ZD}
\title{EUCAST Data}
\format{A data frame with 1476 observations on the following 51 variables:
  \describe{
     \item{\code{Antimicrobial}}{a factor with levels \code{Amikacin} \code{Amoxicillin} \code{Amoxicillin-clavulanic acid (fixed)} \code{Ampicillin} \code{Ampicillin-sulbactam (fixed)} \code{Azithromycin} \code{Aztreonam} \code{Benzylpenicillin} \code{Cefaclor} \code{Cefadroxil} \code{Cefalexin} \code{Cefepime} \code{Cefixime} \code{Cefotaxime} \code{Cefoxitin} \code{Cefpodoxime} \code{Ceftaroline} \code{Ceftazidime} \code{Ceftazidime-avibactam} \code{Ceftibuten} \code{Ceftiofur} \code{Ceftolozane-tazobactam} \code{Ceftriaxone} \code{Cefuroxime} \code{Chloramphenicol} \code{Ciprofloxacin} \code{Clindamycin} \code{Doripenem} \code{Enrofloxacin} \code{Ertapenem} \code{Erythromycin} \code{Florfenicol} \code{Fosfomycin} \code{Fusidic acid} \code{Gentamicin} \code{Imipenem} \code{Levofloxacin} \code{Linezolid} \code{Mecillinam} \code{Meropenem} \code{Minocycline} \code{Moxifloxacin} \code{Mupirocin} \code{Nalidixic acid} \code{Neomycin} \code{Netilmicin} \code{Nitrofurantoin} \code{Nitroxoline} \code{Norfloxacin} \code{Ofloxacin} \code{Oxacillin} \code{Pefloxacin} \code{Piperacillin} \code{Piperacillin-tazobactam} \code{Quinupristin-dalfopristin} \code{Rifampicin} \code{Streptomycin} \code{Teicoplanin} \code{Telithromycin} \code{Temocillin} \code{Tetracycline} \code{Ticarcillin} \code{Ticarcillin-clavulanic acid} \code{Tigecycline} \code{Tobramycin} \code{Trimethoprim} \code{Trimethoprim-sulfamethoxazole} \code{Vancomycin}}
   \item{\code{Bacterium}}{a factor with levels \code{Acinetobacter baumannii} \code{Acinetobacter pittii} \code{Acinetobacter spp} \code{Aerococcus sanguinicola} \code{Aerococcus urinae} \code{Aeromonas spp} \code{Campylobacter coli} \code{Campylobacter jejuni} \code{Campylobacter jejuni ATCC 33560} \code{Citrobacter amalonaticus} \code{Citrobacter braakii} \code{Citrobacter freundii} \code{Citrobacter koseri} \code{Citrobacter spp} \code{Corynebacterium afermentans} \code{Corynebacterium amycolatum} \code{Corynebacterium aurimucosum} \code{Corynebacterium glucuronolyticum} \code{Corynebacterium jeikeium} \code{Corynebacterium propinquum} \code{Corynebacterium pseudodiphtheriticum} \code{Corynebacterium singulare} \code{Corynebacterium spp} \code{Corynebacterium striatum} \code{Corynebacterium urealyticum} \code{Enterobacter asburiae} \code{Enterobacter cloacae} \code{Enterobacter spp} \code{Enterococcus faecalis} \code{Enterococcus faecalis ATCC 29212} \code{Enterococcus faecalis ATCC 51299} \code{Enterococcus faecium} \code{Escherichia coli} \code{Escherichia coli ATCC 25922} \code{Escherichia coli ATCC 35218} \code{Haemophilus influenzae} \code{Haemophilus influenzae ATCC 49247} \code{Haemophilus influenzae ATCC 49766} \code{Haemophilus influenzae NCTC 8468} \code{Haemophilus parainfluenzae} \code{Hafnia alvei} \code{Kingella kingae} \code{Klebsiella aerogenes} \code{Klebsiella oxytoca} \code{Klebsiella pneumoniae} \code{Klebsiella pneumoniae ATCC 700603} \code{Listeria monocytogenes} \code{Mannheimia haemolytica} \code{Moraxella catarrhalis} \code{Morganella morganii} \code{Pasteurella canis} \code{Pasteurella multocida} \code{Proteus mirabilis} \code{Proteus vulgaris} \code{Providencia rettgeri} \code{Pseudomonas aeruginosa} \code{Pseudomonas aeruginosa ATCC 27853} \code{Pseudomonas fluorescens group} \code{Pseudomonas putida group} \code{Pseudomonas spp (non-aeruginosa)} \code{Pseudomonas stutzeri group} \code{Salmonella enteritidis} \code{Salmonella spp} \code{Serratia marcescens} \code{Shigella boydii} \code{Shigella dysenteriae} \code{Shigella flexneri} \code{Shigella sonnei} \code{Shigella spp} \code{Staphylococcus aureus} \code{Staphylococcus aureus ATCC 29213} \code{Staphylococcus aureus MRSA} \code{Staphylococcus aureus MSSA} \code{Staphylococcus aureus NCTC 12493} \code{Staphylococcus capitis} \code{Staphylococcus caprae} \code{Staphylococcus coagulase negative} \code{Staphylococcus cohnii} \code{Staphylococcus epidermidis} \code{Staphylococcus haemolyticus} \code{Staphylococcus hominis} \code{Staphylococcus lugdunensis} \code{Staphylococcus pettenkoferi} \code{Staphylococcus pseudintermedius} \code{Staphylococcus saprophyticus} \code{Staphylococcus warneri} \code{Stenotrophomonas maltophilia} \code{Streptococcus agalactiae} \code{Streptococcus anginosus} \code{Streptococcus bovis} \code{Streptococcus dysgalactiae} \code{Streptococcus gordonii} \code{Streptococcus group C} \code{Streptococcus group G} \code{Streptococcus milleri} \code{Streptococcus mitis} \code{Streptococcus oralis} \code{Streptococcus parasanguis} \code{Streptococcus pneumoniae} \code{Streptococcus pneumoniae ATCC 49619} \code{Streptococcus pyogenes} \code{Streptococcus pyogenes CCUG 25571} \code{Streptococcus salivarius} \code{Streptococcus sanguinis} \code{Streptococcus viridans} \code{Streptococcus, viridans group} \code{Yersinia enterocolitica}}
   \item{\code{DiskContent}}{a numeric vector}
    \item{\code{Z6}}{a numeric vector}
    \item{\code{Z7}}{a numeric vector}
    \item{\code{Z8}}{a numeric vector}
    \item{\code{Z9}}{a numeric vector}
    \item{\code{Z10}}{a numeric vector}
    \item{\code{Z11}}{a numeric vector}
    \item{\code{Z12}}{a numeric vector}
    \item{\code{Z13}}{a numeric vector}
    \item{\code{Z14}}{a numeric vector}
    \item{\code{Z15}}{a numeric vector}
    \item{\code{Z16}}{a numeric vector}
    \item{\code{Z17}}{a numeric vector}
    \item{\code{Z18}}{a numeric vector}
    \item{\code{Z19}}{a numeric vector}
    \item{\code{Z20}}{a numeric vector}
    \item{\code{Z21}}{a numeric vector}
    \item{\code{Z22}}{a numeric vector}
    \item{\code{Z23}}{a numeric vector}
    \item{\code{Z24}}{a numeric vector}
    \item{\code{Z25}}{a numeric vector}
    \item{\code{Z26}}{a numeric vector}
    \item{\code{Z27}}{a numeric vector}
    \item{\code{Z28}}{a numeric vector}
    \item{\code{Z29}}{a numeric vector}
    \item{\code{Z30}}{a numeric vector}
    \item{\code{Z31}}{a numeric vector}
    \item{\code{Z32}}{a numeric vector}
    \item{\code{Z33}}{a numeric vector}
    \item{\code{Z34}}{a numeric vector}
    \item{\code{Z35}}{a numeric vector}
    \item{\code{Z36}}{a numeric vector}
    \item{\code{Z37}}{a numeric vector}
    \item{\code{Z38}}{a numeric vector}
    \item{\code{Z39}}{a numeric vector}
    \item{\code{Z40}}{a numeric vector}
    \item{\code{Z41}}{a numeric vector}
    \item{\code{Z42}}{a numeric vector}
    \item{\code{Z43}}{a numeric vector}
    \item{\code{Z44}}{a numeric vector}
    \item{\code{Z45}}{a numeric vector}
    \item{\code{Z46}}{a numeric vector}
    \item{\code{Z47}}{a numeric vector}
    \item{\code{Z48}}{a numeric vector}
    \item{\code{Z49}}{a numeric vector}
    \item{\code{Z50}}{a numeric vector}
    \item{\code{ECOFF}}{a numeric vector}
    \item{\code{Distributions}}{a numeric vector}
    \item{\code{Observations}}{a numeric vector}
  }}
\source{
The European Committee on Antimicrobial Susceptibility Testing. Antimicrobial wild type distributions of microorganisms, 2019 \url{https://mic.eucast.org/Eucast2/SearchController/search.jsp?action=init}
}
\usage{
ZD
}
\description{
A data set describing the resistance of bacteria to antimicrobials. ZD measures how far the antimicrobial was able to keep the bacterium off. Observations with diameter 6mm are resistant.
}
\keyword{datasets}
